mke2fs: add new -x filter option to exclude files from being copied
Multiple -x options can occur on the command line. Each is a glob style
filter that is matched against names of files being copied into the
filesystem using the -d option.

If the filter starts with '/', the pattern is matched against the full
path relative to the root of the new filesystem. Otherwise the pattern is
matched against file names in any directory.

This is useful to avoid adding unwanted files when building a system image
from a root-directory. For example:

    mke2fs -d rootfs -x '/dev/*' -x '/boot/*' -x '/tmp/*' -x '*.log' out.img

This would leave an empty directory (used as a mount point) for /boot, /dev,
and /tmp but not include any files within them. It would also not include
files with names matching '*.log' in any directory.