mke2fs: add support for SOURCE_DATE_EPOCH when copying files with -d
When copying files into a new filesystem using the -d option, the
timestamps in the inodes are set by using the times read by stat from the
source file.  In order to make a reproducible filesystem image, all the
timestamps must be deterministic. Unfortunately, the crtime (birth), and
ctimes in the source directory cannot be controlled, e.g. with touch. This
means if files in the source directory were ever re-created, there is no
way to reset those timestamps and reproduce the same filesystem image.

When the environment variable SOURCE_DATE_EPOCH is set, all the timestamps
in the inodes of copied files will be set to the value of the variable
instead of the times read by stat. In addition, the fs->now time will also
be set to SOURCE_DATE_EPOCH which is used for any other inode timestamps
in the filesystem and the root directory itself.

This is a superset of the functionality of the existing environment
variable E2FSPROGS_FAKE_TIME which does not fake the times of the copied
files and does not allow for a fake time of zero.