misc: fix strncpy warnings in fuse2fs.c and e4defrag.c
This fixes spurious GCC warning:

warning: ‘strncpy’ specified bound depends on the length of the source
argument [-Wstringop-overflow=]

Calling strncpy with N set to the length of the source string does not
protect the destination from overflow. Calling strncpy(d, s, strlen(s)+1)
is just equivalent to strcpy(). Calling strncpy(d, s, strlen(s)) is
dangerous because it doesn't copy the null terminator.

In fuse2fs, strncpy(d, s, strlen(s)) was not copying the terminator and
relied on the buffer being zeroed in advance. It can safely be replaced
with strcpy() because the destination buffer size was computed dynamically
to account for the terminators and string lengths.

In e4defrag, one destination buffer length was dynamically allocated and
is safe and equivalent to strcpy(). The other two cases have a fixed size
destination.